Add a box for playlist
A listbox widget was used to display a playlist.An add and delete button
were also introduced to add and remove audio files from the playlist.
The playlist is not yet fully funcitonal though.
I hope i will be sorting out current problems and add previous and next
buttons in upcoming days.